Ea-4466 Fix allow headers HEAD value, add HeadersTestIT
